Use system-dependent, world-accessible common file location
This commit updates GlobalFilePortLocker to create/reuse the
port lock file in a system-dependent, generally accessible
common file location (as supported by
org.terracotta.utilities.test.io.CommonFiles.createCommonAppFile)
instead of relying on `java.io.tmpdir` -- which can be
user-specific.  On *NIX systems, the location used is `/var/tmp`
(in compliance with the Filesystem Hierarchy Standard).  On
Windows systems, this is the directory assigned to the
`CommonApplicationData` special folder.