Introduce replacement port management -- PortManager
This commit introduces a replacement for PortLocker, PortManager,
that merges functionality from both PortChooser (out of
terracotta-core) and PortLocker in an effort to improve TCP port
reservation stability.

A major difference between PortManager and PortLocker is the
removal of methods returning a reservation of a consecutive
sequence of ports.  This is replaced with returning a list
of reserved ports -- each randomly chosen.  PortManager, like
PortChooser, avoids reservations from the ephemeral (dynamic)
pool to avoid races between reserving and using a port and
dynamic socket requests by other applications (or test clients).