Correct Files.relocate handling of relative and existing targets
This commit corrects the behavior of Files.relocate in two cases:

. where target is relative
. where target exists and StandardOptions.REPLACE_EXISTING
    is not specified

When the target is relative, relocate was improperly resolving
the target as a sibling of the source.  This the documented
behavior of rename but is not the expected behavior for
relocate.  Relocate now resolves against the default directory
for the file system (generally user.dir).

When the target exists and REPLACE_EXISTING is not specified,
the expected behavior is to receive an exception -- generally a
FileAlreadyExistsException.  Relocate, when REPLACE_EXISTING is
not specified and under some conditions, was replacing an existing
target.  It now throws a FileAlreadyExistsException if the
target exists and REPLACE_EXISTING is not specified.