Add support for Files.createFile
This commit adds a Files.createFile method that handles the
AccessDeniedException thrown from java.nio.file.Files.createFile
when the target file is just marked for deletion and not yet
deleted (Windows).

To support testing of this method, InterprocessCyclicBarrier is added
to permit test coordination among Java processes.

This commit resolves #50.