Move to 3-position snapshot version
This commit updates the snapshot version numbering to include 3 positions
instead of the previous 2.  This corrects an issue with version resolution
when ranges are involved -- 0.0-SNAPSHOT is, unfortunately, less than 0.0.x
for any value of x in both the Maven and Gradle version comparison schemes.
Moving to a 3 position snapshot, e.g. 0.0.13-SNAPSHOT avoids this problem.