Use version range for logging components
This commit shifts to the use of version ranges for SLF4J and logback.
SLF4J is capped at 1.7.9999 to avoid picking up 1.8.0-betaX releases
which will not work with released versions of Logback before 1.3.x.
The minimum versions of Logback is 1.2.11 which requires a minumum
version of SLF4J of 1.7.32.

While developing this update, it was observed that the SLF4J version
must "align" with the Logback version for Logback to be properly
initialized.  With this observation, tests for ConsoleAppenderCapture
are changed to terminate with a more informative assertion to avoid
an obscure test failure.

The configuration for the enforcer plugin is lifted out of an
execution configuration to permit its use as a top-level task.