Apply exclusion to correct dependency convergence failure
Module terracotta-utilities-port-chooser fails dependency convergence
verification because dependency convergence analysis fixes the version
of SLF4J used in terracotta-utilities-tools at the minimum version of
the range even though the dependency actually resolves the latest
available in the range.  The terracotta-utilities-port-chooser POM is
changed to exclude the SLF4J transitive dependency of
terracotta-utilities-tools in deference to the direct dependency.