Limit query done by port release check to port being released
Before this commit, the PortManager port release check queried details
about all open ports and filtered the results to the port being released.
This proved quite slow in some circumstances so NetStat is changed to
add a method that, on Linux and Windows, queries only the port being
checked.  (The port query method used for Mac OS X does not support
querying a single port.)  PortManager is changed to use this
single-port method when performing the release check.